+++ arch/arc/core/cpu_idle.S
@@ -35,9 +35,9 @@ SECTION_VAR(BSS, k_cpu_sleep_mode)
 
 SECTION_FUNC(TEXT, k_cpu_idle)
 
-#ifdef CONFIG_KERNEL_EVENT_LOGGER_SLEEP
+#ifdef CONFIG_TRACING
 	push_s blink
-	jl    _sys_k_event_logger_enter_sleep
+	jl    sys_trace_idle
 	pop_s blink
 #endif
 
@@ -56,9 +56,9 @@ SECTION_FUNC(TEXT, k_cpu_idle)
  */
 SECTION_FUNC(TEXT, k_cpu_atomic_idle)
 
-#ifdef CONFIG_KERNEL_EVENT_LOGGER_SLEEP
+#ifdef CONFIG_TRACING
 	push_s blink
-	jl    _sys_k_event_logger_enter_sleep
+	jl    sys_trace_idle
 	pop_s blink
 #endif
