+++ arch/arc/core/isr_wrapper.S
@@ -332,29 +332,14 @@ rirq_path:
 	j_s [r2]
 #endif
 
-#ifdef CONFIG_KERNEL_EVENT_LOGGER_SLEEP
-GTEXT(_sys_k_event_logger_exit_sleep)
-
-.macro log_sleep_k_event
-	clri r0 /* do not interrupt event logger operations */
-	push_s r0
-	push_s blink
-	jl _sys_k_event_logger_exit_sleep
-	pop_s blink
-	pop_s r0
-	seti r0
-.endm
-#else
-	#define log_sleep_k_event
-#endif
-#if defined(CONFIG_KERNEL_EVENT_LOGGER_INTERRUPT)
-GTEXT(_sys_k_event_logger_interrupt)
+#if defined(CONFIG_TRACING)
+GTEXT(sys_trace_isr_enter)
 
 .macro log_interrupt_k_event
 	clri r0 /* do not interrupt event logger operations */
 	push_s r0
 	push_s blink
-	jl _sys_k_event_logger_interrupt
+	jl sys_trace_isr_enter
 	pop_s blink
 	pop_s r0
 	seti r0
@@ -395,7 +380,6 @@ SECTION_FUNC(TEXT, _isr_demux)
 	/* r0 is available to be stomped here, and exit_tickless_idle uses it */
 	exit_tickless_idle
 	log_interrupt_k_event
-	log_sleep_k_event
 
 	lr r0, [_ARC_V2_ICAUSE]
 	/* handle software triggered interrupt */
