+++ arch/nios2/include/tracing_arch.h
@@ -1,16 +1,16 @@
 /*
- * Copyright (c) 2016 Jean-Paul Etienne <fractalclone@gmail.com>
+ * Copyright (c) 2016 Intel Corporation
  *
  * SPDX-License-Identifier: Apache-2.0
  */
 
 /**
  * @file
- * @brief Kernel event logger support for RISCV32
+ * @brief Kernel event logger support for Nios II
  */
 
-#ifndef __KERNEL_EVENT_LOGGER_ARCH_H__
-#define __KERNEL_EVENT_LOGGER_ARCH_H__
+#ifndef __KERNEL_TRACING_H__
+#define __KERNEL_TRACING_H__
 
 #include <arch/cpu.h>
 
@@ -22,23 +22,20 @@ extern "C" {
  * @brief Get the identification of the current interrupt.
  *
  * This routine obtain the key of the interrupt that is currently processed
- * if it is called from an IRQ context.
+ * if it is called from a ISR context.
  *
  * @return The key of the interrupt that is currently being processed.
  */
 static inline int _sys_current_irq_key_get(void)
 {
-	u32_t mcause;
+	u32_t ipending;
 
-	__asm__ volatile("csrr %0, mcause" : "=r" (mcause));
-
-	mcause &= SOC_MCAUSE_EXP_MASK;
-
-	return mcause;
+	ipending = _nios2_creg_read(NIOS2_CR_IPENDING);
+	return find_lsb_set(ipending) - 1;
 }
 
 #ifdef __cplusplus
 }
 #endif
 
-#endif /* __KERNEL_EVENT_LOGGER_ARCH_H__ */
+#endif /* __KERNEL_TRACING_H__ */
