+++ arch/riscv32/core/isr.S
@@ -21,16 +21,9 @@ GTEXT(_k_neg_eagain)
 GTEXT(_is_next_thread_current)
 GTEXT(_get_next_ready_thread)
 
-#ifdef CONFIG_KERNEL_EVENT_LOGGER_CONTEXT_SWITCH
-GTEXT(_sys_k_event_logger_context_switch)
-#endif
-
-#ifdef CONFIG_KERNEL_EVENT_LOGGER_SLEEP
-GTEXT(_sys_k_event_logger_exit_sleep)
-#endif
-
-#ifdef CONFIG_KERNEL_EVENT_LOGGER_INTERRUPT
-GTEXT(_sys_k_event_logger_interrupt)
+#ifdef CONFIG_TRACING
+GTEXT(sys_trace_thread_switched_in)
+GTEXT(sys_trace_isr_enter)
 #endif
 
 #ifdef CONFIG_IRQ_OFFLOAD
@@ -231,12 +224,8 @@ on_irq_stack:
 	tail _irq_do_offload
 
 call_irq:
-#ifdef CONFIG_KERNEL_EVENT_LOGGER_SLEEP
-	call _sys_k_event_logger_exit_sleep
-#endif
-
-#ifdef CONFIG_KERNEL_EVENT_LOGGER_INTERRUPT
-	call _sys_k_event_logger_interrupt
+#ifdef CONFIG_TRACING
+	call sys_trace_isr_enter
 #endif
 
 	/* Get IRQ causing interrupt */
@@ -310,9 +299,9 @@ reschedule:
 #if CONFIG_TIMESLICING
 	call _update_time_slice_before_swap
 #endif
-#if CONFIG_KERNEL_EVENT_LOGGER_CONTEXT_SWITCH
-	call _sys_k_event_logger_context_switch
-#endif /* CONFIG_KERNEL_EVENT_LOGGER_CONTEXT_SWITCH */
+#if CONFIG_TRACING
+	call sys_trace_thread_switched_in
+#endif
 	/* Get reference to _kernel */
 	la t0, _kernel
