+++ arch/riscv32/soc/riscv-privilege/common/idle.c
@@ -9,13 +9,11 @@
 #include <irq.h>
 #include <soc.h>
 
-#include <logging/kernel_event_logger.h>
+#include <tracing.h>
 
 static ALWAYS_INLINE void riscv_idle(unsigned int key)
 {
-#ifdef CONFIG_KERNEL_EVENT_LOGGER_SLEEP
-	_sys_k_event_logger_enter_sleep();
-#endif
+	sys_trace_idle();
 	/* unlock interrupts */
 	irq_unlock(key);
