+++ arch/xtensa/include/tracing_arch.h
@@ -1,18 +1,16 @@
 /*
- * Copyright (c) 2015 Intel Corporation
- *
+ * Copyright (c) 2016 Cadence Design Systems, Inc.
  * SPDX-License-Identifier: Apache-2.0
  */
 
 /**
  * @file
- * @brief Kernel event logger support for x86
+ * @brief Kernel event logger support for Xtensa
  */
 
-#ifndef __KERNEL_EVENT_LOGGER_ARCH_H__
-#define __KERNEL_EVENT_LOGGER_ARCH_H__
+#ifndef __KERNEL_TRACING_H__
+#define __KERNEL_TRACING_H__
 
-#include <arch/x86/irq_controller.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -28,11 +26,11 @@ extern "C" {
  */
 static inline int _sys_current_irq_key_get(void)
 {
-	return _irq_controller_isr_vector_get();
+	return 0;
 }
 
 #ifdef __cplusplus
 }
 #endif
 
-#endif /* __KERNEL_EVENT_LOGGER_ARCH_H__ */
+#endif /* __KERNEL_TRACING_H__ */
