+++ doc/subsystems/logging/index.rst
@@ -5,5 +5,4 @@ Logging
    :maxdepth: 1
 
    system_log
-   kernel_event_logger
    logger
