+++ doc/zephyr.doxyfile.in
@@ -1969,9 +1969,6 @@ INCLUDE_FILE_PATTERNS  =
 # This tag requires that the tag ENABLE_PREPROCESSING is set to YES.
 
 PREDEFINED             = "CONFIG_SYS_CLOCK_EXISTS=y" \
-                         "CONFIG_KERNEL_EVENT_LOGGER=y" \
-                         "CONFIG_KERNEL_EVENT_LOGGER_DYNAMIC=y" \
-                         "CONFIG_KERNEL_EVENT_LOGGER_CUSTOM_TIMESTAMP=y" \
                          "CONFIG_THREAD_MONITOR=y" \
                          "CONFIG_THREAD_CUSTOM_DATA=y" \
                          "CONFIG_ERRNO=y" \
