--- kernel/Kconfig.event_logger
@@ -1,76 +1,0 @@
-#
-# Copyright (c) 2014-2015 Wind River Systems, Inc.
-#
-# SPDX-License-Identifier: Apache-2.0
-#
-
-menuconfig KERNEL_EVENT_LOGGER
-	bool "Enable kernel event logger features"
-	select RING_BUFFER
-	help
-	  This feature enables the usage of the profiling logger. Provides the
-	  logging of sleep events (either entering or leaving low power conditions),
-	  context switch events, interrupt events, boot events and a method to
-	  collect these event messages.
-
-if KERNEL_EVENT_LOGGER
-config KERNEL_EVENT_LOGGER_BUFFER_SIZE
-	int "Kernel event logger buffer size"
-	default 128
-	help
-	  Buffer size in 32-bit words.
-
-config KERNEL_EVENT_LOGGER_DYNAMIC
-	bool "Kernel event logger dynamic enabling"
-	help
-	  If enabled, kernel event logger is not logging any data to the ring buffer
-	  It is up to the application to set the appropriate flags to enable/disable the
-	  logging of each event type.
-
-config KERNEL_EVENT_LOGGER_CUSTOM_TIMESTAMP
-	bool "Kernel event logger custom timestamp"
-	help
-	  This flag enables the possibility to set the timer function to be used to
-	  populate kernel event logger timestamp. This has to be done at runtime by
-	  calling sys_k_event_logger_set_timer and providing the function callback.
-
-menu "Kernel event logging points"
-
-config KERNEL_EVENT_LOGGER_CONTEXT_SWITCH
-	bool "Context switch event logging point"
-	help
-	  Enable the context switch event messages.
-
-config KERNEL_EVENT_LOGGER_INTERRUPT
-	bool "Interrupt event logging point"
-	help
-	  Enable interrupt event messages. These messages provide the following
-	  information: The time when interrupts occur.
-
-config KERNEL_EVENT_LOGGER_SLEEP
-	bool "Sleep event logging point"
-	depends on SYS_POWER_MANAGEMENT
-	help
-	  Enable low power condition event messages. These messages provide the
-	  following information:
-
-		- When the CPU went to sleep mode.
-		- When the CPU woke up.
-		- The ID of the interrupt that woke the CPU up.
-
-config KERNEL_EVENT_LOGGER_THREAD
-	bool "Thread event logging point"
-	help
-	  Enable thread event messages. These messages provide the following
-	  information:
-
-		- When threads are marked as ready to be executed.
-		- When threads are marked as pending.
-		- When threads end.
-
-	  Context switching messages should be enabled by enabling the
-	  CONFIG_KERNEL_EVENT_LOGGER_CONTEXT_SWITCH setting instead.
-
-endmenu
-
-endif
