+++ kernel/thread.c
@@ -560,9 +560,6 @@ void _k_thread_single_abort(struct k_thread *thread)
 	thread->base.thread_state |= _THREAD_DEAD;
 
 	sys_trace_thread_abort(thread);
-#ifdef CONFIG_KERNEL_EVENT_LOGGER_THREAD
-	_sys_k_event_logger_thread_exit(thread);
-#endif
 
 #ifdef CONFIG_USERSPACE
 	/* Clear initailized state so that this thread object may be re-used
