--- samples/net/zperf/prj_frdm_k64f_prof.conf
@@ -1,37 +1,0 @@
-CONFIG_NETWORKING=y
-CONFIG_NET_LOG=y
-CONFIG_NET_IPV6=y
-CONFIG_NET_IPV4=y
-CONFIG_NET_DHCPV4=y
-CONFIG_NET_UDP=y
-CONFIG_NET_TCP=y
-CONFIG_NET_STATISTICS=y
-
-CONFIG_NET_PKT_RX_COUNT=14
-CONFIG_NET_PKT_TX_COUNT=14
-CONFIG_NET_BUF_RX_COUNT=30
-CONFIG_NET_BUF_TX_COUNT=30
-CONFIG_NET_BUF_DATA_SIZE=128
-CONFIG_NET_IF_UNICAST_IPV6_ADDR_COUNT=5
-CONFIG_NET_IF_MCAST_IPV6_ADDR_COUNT=5
-CONFIG_NET_IF_UNICAST_IPV4_ADDR_COUNT=2
-CONFIG_NET_MAX_CONTEXTS=10
-
-CONFIG_INIT_STACKS=y
-CONFIG_TEST_RANDOM_GENERATOR=y
-CONFIG_SYS_LOG_SHOW_COLOR=y
-CONFIG_PRINTK=y
-
-CONFIG_NET_L2_ETHERNET=y
-
-CONFIG_NET_SHELL=y
-
-#
-# Profiler
-#
-CONFIG_RING_BUFFER=y
-CONFIG_KERNEL_EVENT_LOGGER=y
-CONFIG_KERNEL_EVENT_LOGGER_BUFFER_SIZE=10000
-CONFIG_KERNEL_EVENT_LOGGER_CONTEXT_SWITCH=y
-CONFIG_KERNEL_EVENT_LOGGER_INTERRUPT=y
-CONFIG_KERNEL_EVENT_LOGGER_DYNAMIC=y
